+++ core/common/src/main/kotlin/com/kolown/porring/core/common/component/ShimmerEffect.kt
@@ -13,11 +13,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.composed
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.unit.IntSize
-import com.kolown.designsystem.ui.theme.ShimmerDarkGray
-import com.kolown.designsystem.ui.theme.ShimmerLightGray
+import com.kolown.porring.core.designsystem.ui.theme.ShimmerDarkGray
+import com.kolown.porring.core.designsystem.ui.theme.ShimmerLightGray
 
 fun Modifier.shimmerEffect(): Modifier = composed {
     var size by remember {
