+++ core/designsystem/src/main/java/com/kolown/porring/core/designsystem/component/PorringCenterAlignTopAppBar.kt
@@ -1,11 +1,13 @@
-package com.kolown.designsystem.component
+package com.kolown.porring.core.designsystem.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -15,7 +17,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 
 @Composable
-fun PorringTopAppBar(
+fun PorringCenterAlignTopAppBar(
     title: String? = null,
     navigationIcon: (@Composable () -> Unit)? = null,
     trailingIcon: (@Composable () -> Unit)? = null,
@@ -28,26 +30,27 @@ fun PorringTopAppBar(
             .background(Color.Transparent)
             .padding(4.dp),
         verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = if (navigationIcon == null && trailingIcon != null) Arrangement.End else Arrangement.SpaceBetween
+        horizontalArrangement = Arrangement.SpaceBetween
     ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            navigationIcon?.let {
-                navigationIcon()
-            }
 
-            title?.let {
-                Text(
-                    style = MaterialTheme.typography.titleLarge,
-                    text = title,
-                    modifier = Modifier.padding(start = 16.dp)
-                )
-            }
+        if (navigationIcon == null ) {
+            Spacer(Modifier.size(48.dp))
+        } else {
+            navigationIcon()
         }
 
 
-        trailingIcon?.let {
+        title?.let {
+            Text(
+                style = MaterialTheme.typography.titleLarge,
+                text = title,
+                modifier = Modifier
+            )
+        }
+
+        if (trailingIcon == null) {
+            Spacer(Modifier.size(48.dp))
+        } else {
             trailingIcon()
         }
     }
