+++ feature/home/src/main/java/com/kolown/porring/feature/home/HomeScreen.kt
@@ -38,7 +38,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.kolown.designsystem.ui.theme.Primary
+import com.kolown.porring.core.designsystem.ui.theme.Primary
 import com.kolown.porring.feature.home.component.RandomImageList
 import com.kolown.model.PostContentModel
 import com.kolown.model.Reactions
