+++ feature/home/src/main/java/com/kolown/porring/feature/home/component/RandomImageList.kt
@@ -40,7 +40,7 @@ import androidx.compose.ui.unit.dp
 import com.kolown.porring.core.common.component.CoilImage
 import com.kolown.porring.core.common.component.FollowDialog
 import com.kolown.porring.core.common.component.LocalSnackBarBridge
-import com.kolown.designsystem.ui.theme.Primary
+import com.kolown.porring.core.designsystem.ui.theme.Primary
 import com.kolown.model.PostContentModel
 import com.kolown.model.Reactions
 import com.kolown.model.SnackBarEvent
