+++ feature/join/src/main/java/com/kolown/porring/feature/join/JoinScreen.kt
@@ -44,12 +44,12 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.google.firebase.auth.FirebaseAuthUserCollisionException
 import com.kolown.porring.core.common.component.LocalSnackBarBridge
-import com.kolown.designsystem.component.PorringIconButton
-import com.kolown.designsystem.component.PorringTextField
-import com.kolown.designsystem.component.PorringTopAppBar
-import com.kolown.designsystem.ui.theme.Error
-import com.kolown.designsystem.ui.theme.Primary
-import com.kolown.designsystem.ui.theme.PrimaryUnActive
+import com.kolown.porring.core.designsystem.component.PorringIconButton
+import com.kolown.porring.core.designsystem.component.PorringTextField
+import com.kolown.porring.core.designsystem.component.PorringTopAppBar
+import com.kolown.porring.core.designsystem.ui.theme.Error
+import com.kolown.porring.core.designsystem.ui.theme.Primary
+import com.kolown.porring.core.designsystem.ui.theme.PrimaryUnActive
 import com.kolown.model.UiState
 import com.kolown.navigation.Route
