+++ feature/login/src/main/java/com/kolown/porring/feature/login/LoginScreen.kt
@@ -59,12 +59,12 @@ import androidx.lifecycle.lifecycleScope
 import com.google.firebase.FirebaseNetworkException
 import com.google.firebase.auth.FirebaseAuthInvalidCredentialsException
 import com.kolown.porring.core.common.component.LocalSnackBarBridge
-import com.kolown.designsystem.R
-import com.kolown.designsystem.component.PorringIconButton
-import com.kolown.designsystem.component.PorringTextField
-import com.kolown.designsystem.component.PorringTopAppBar
-import com.kolown.designsystem.ui.theme.Primary
-import com.kolown.designsystem.ui.theme.PrimaryUnActive
+import com.kolown.porring.core.designsystem.R
+import com.kolown.porring.core.designsystem.component.PorringIconButton
+import com.kolown.porring.core.designsystem.component.PorringTextField
+import com.kolown.porring.core.designsystem.component.PorringTopAppBar
+import com.kolown.porring.core.designsystem.ui.theme.Primary
+import com.kolown.porring.core.designsystem.ui.theme.PrimaryUnActive
 import com.kolown.porring.feature.login.R.drawable
 import com.kolown.porring.feature.login.R.string
 import com.kolown.porring.feature.login.component.LoginButtonGroup
