+++ feature/main/src/main/java/com/kolown/porring/feature/main/MainActivity.kt
@@ -14,7 +14,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import com.kolown.porring.core.common.component.LocalSnackBarBridge
 import com.kolown.porring.core.common.component.SnackBarBridge
-import com.kolown.designsystem.ui.theme.PorringTheme
+import com.kolown.porring.core.designsystem.ui.theme.PorringTheme
 import com.kolown.porring.feature.main.component.NotAvailableVersionScreen
 import com.kolown.porring.feature.main.navigation.MainNavigator
 import com.kolown.porring.feature.main.navigation.rememberMainNavigator
