+++ feature/main/src/main/java/com/kolown/porring/feature/main/component/NotAvailableVersionScreen.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import com.kolown.porring.core.common.R
-import com.kolown.designsystem.ui.theme.Primary
+import com.kolown.porring.core.designsystem.ui.theme.Primary
 
 @Composable
 fun NotAvailableVersionScreen(
