+++ feature/setting/src/main/java/com/kolown/porring/feature/setting/component/MenuDivider.kt
@@ -7,7 +7,7 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.kolown.designsystem.ui.theme.Surface
+import com.kolown.porring.core.designsystem.ui.theme.Surface
 
 @Composable
 internal fun MenuDivider() {
