+++ feature/upload/src/main/java/com/kolown/porring/feature/upload/component/TagChipGroup.kt
@@ -23,9 +23,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.dp
-import com.kolown.designsystem.component.PorringIconButton
-import com.kolown.designsystem.component.PorringInputChip
-import com.kolown.designsystem.ui.theme.Primary
+import com.kolown.porring.core.designsystem.component.PorringIconButton
+import com.kolown.porring.core.designsystem.component.PorringInputChip
+import com.kolown.porring.core.designsystem.ui.theme.Primary
 import com.kolown.porring.feature.upload.R
 
 @OptIn(ExperimentalLayoutApi::class)
